chore(token-price): dhedgepoolaggregator cast

+++ contracts/assets/DHedgePoolAggregator.sol
@@ -46,7 +46,7 @@ contract DHedgePoolAggregator is IAggregatorV3Interface {
   {
     uint256 tokenPrice = IPoolLogic(poolLogic).tokenPrice();
     // adjust decimals -> 8
-    int256 answer = tokenPrice == 0 ? 0 : tokenPrice.div(10**10);
+    int256 answer = tokenPrice == 0 ? 0 : int256(tokenPrice.div(10**10));
     // we don't need roundId, startedAt and answeredInRound
     return (0, answer, 0, block.timestamp, 0);
   }
